requirements.txt
accelerate== 0.27.2
appnope== 0.1.4
asttokens== 2.4.1
backcall== 0.2.0
certifi== 2024.2.2
charset-normalizer== 3.3.2
comm== 0.2.1
debugpy== 1.6.7
decorator== 5.1.1
executing== 2.0.1
filelock== 3.13.1
fsspec== 2024.2.0
huggingface== 0.0.1
huggingface-hub== 0.21.4
idna== 3.6
importlib_metadata== 7.0.2
ipykernel== 6.29.3
ipython== 8.12.0
jedi== 0.19.1
Jinja2== 3.1.3
jupyter_client== 8.6.0
jupyter_core== 5.7.1
MarkupSafe== 2.1.5
matplotlib-inline== 0.1.6
mpmath== 1.3.0
nest-asyncio== 1.6.0
networkx== 3.2.1
nibabel== 4.0.2
numpy== 1.26.4
packaging== 23.2
parso== 0.8.3
peft== 0.9.0
pexpect== 4.9.0
pickleshare== 0.7.5
pip== 23.3.1
platformdirs== 4.2.0
prompt-toolkit== 3.0.42
protobuf== 4.25.3
psutil== 5.9.8
ptyprocess== 0.7.0
pure-eval== 0.2.2
Pygments== 2.17.2
python-dateutil== 2.9.0
python-dotenv== 1.0.1
PyYAML== 6.0.1
pyzmq== 24.0.1
regex== 2023.12.25
requests== 2.31.0
safetensors== 0.4.2
sentencepiece== 0.2.0
setuptools== 68.2.2
six== 1.16.0
stack-data== 0.6.2
sympy== 1.12
tokenizers== 0.15.2
torch== 2.2.1
tornado== 6.4
tqdm== 4.66.2
traitlets== 5.14.1
transformers== 4.38.2
typing_extensions== 4.10.0
urllib3== 2.2.1
wcwidth== 0.2.13
wheel== 0.41.2
zipp== 3.17.0
